Microsoft-Webpage-Clone
In this project, I looked at Microsoft's own page, designed and coded the pictures, buttons and everything necessary to be compatible with the computer screen, tablet screen and phone screen, and finally a sidebar was coded for looking at the phone. I completed this project with the video I watched, through the files seen in all other sections. to reach.